# *: Optimize struct memory usage by adjust field order