--- a/expression/aggregation/concat.go
+++ b/expression/aggregation/concat.go
@@ -28,8 +28,8 @@ import (
 type concatFunction struct {
 	aggFunction
 	separator string
-	sepInited bool
 	maxLen    uint64
+	sepInited bool
 	// truncated according to MySQL, a 'group_concat' function generates exactly one 'truncated' warning during its life time, no matter
 	// how many group actually truncated. 'truncated' acts as a sentinel to indicate whether this warning has already been
 	// generated.