--- a/planner/core/common_plans.go
+++ b/planner/core/common_plans.go
@@ -371,11 +371,11 @@ type Show struct {
 	Table       *ast.TableName  // Used for showing columns.
 	Column      *ast.ColumnName // Used for `desc table column`.
 	IndexName   model.CIStr
-	Flag        int // Some flag parsed from sql, such as FULL.
-	Full        bool
+	Flag        int                  // Some flag parsed from sql, such as FULL.
 	User        *auth.UserIdentity   // Used for show grants.
 	Roles       []*auth.RoleIdentity // Used for show grants.
-	IfNotExists bool                 // Used for `show create database if not exists`
+	Full        bool
+	IfNotExists bool // Used for `show create database if not exists`
 
 	GlobalScope bool // Used by show variables
 }