--- a/planner/core/physical_plans.go
+++ b/planner/core/physical_plans.go
@@ -98,11 +98,6 @@ type PhysicalIndexScan struct {
 	Ranges     []*ranger.Range
 	Columns    []*model.ColumnInfo
 	DBName     model.CIStr
-	Desc       bool
-	KeepOrder  bool
-	// DoubleRead means if the index executor will read kv two times.
-	// If the query requires the columns that don't belong to index, DoubleRead will be true.
-	DoubleRead bool
 
 	TableAsName *model.CIStr
 
@@ -117,10 +112,16 @@ type PhysicalIndexScan struct {
 	rangeInfo string
 
 	// The index scan may be on a partition.
-	isPartition     bool
 	physicalTableID int64
 
 	GenExprs map[model.TableColumnID]expression.Expression
+
+	isPartition bool
+	Desc        bool
+	KeepOrder   bool
+	// DoubleRead means if the index executor will read kv two times.
+	// If the query requires the columns that don't belong to index, DoubleRead will be true.
+	DoubleRead bool
 }
 
 // PhysicalMemTable reads memory table.
@@ -144,27 +145,27 @@ type PhysicalTableScan struct {
 	Table   *model.TableInfo
 	Columns []*model.ColumnInfo
 	DBName  model.CIStr
-	Desc    bool
 	Ranges  []*ranger.Range
 	pkCol   *expression.Column
 
 	TableAsName *model.CIStr
 
-	// KeepOrder is true, if sort data by scanning pkcol,
-	KeepOrder bool
-
 	// Hist is the histogram when the query was issued.
 	// It is used for query feedback.
 	Hist *statistics.Histogram
 
-	// The table scan may be a partition, rather than a real table.
-	isPartition     bool
 	physicalTableID int64
 
 	rangeDecidedBy []*expression.Column
 
 	// HandleIdx is the index of handle, which is only used for admin check table.
 	HandleIdx int
+
+	// The table scan may be a partition, rather than a real table.
+	isPartition bool
+	// KeepOrder is true, if sort data by scanning pkcol,
+	KeepOrder bool
+	Desc      bool
 }
 
 // IsPartition returns true and partition ID if it's actually a partition.