--- a/server/conn.go
+++ b/server/conn.go
@@ -139,7 +139,6 @@ type clientConn struct {
 	server       *Server           // a reference of server instance.
 	capability   uint32            // client capability affects the way server handles client request.
 	connectionID uint32            // atomically allocated by a global variable, unique in process scope.
-	collation    uint8             // collation used by client, may be different from the collation used by database.
 	user         string            // user of the client.
 	dbname       string            // default database name.
 	salt         []byte            // random bytes used for authentication.
@@ -147,10 +146,11 @@ type clientConn struct {
 	lastCmd      string            // latest sql query string, currently used for logging error.
 	ctx          QueryCtx          // an interface to execute sql statements.
 	attrs        map[string]string // attributes parsed from client handshake response, not used for now.
-	status       int32             // dispatching/reading/shutdown/waitshutdown
 	peerHost     string            // peer host
 	peerPort     string            // peer port
+	status       int32             // dispatching/reading/shutdown/waitshutdown
 	lastCode     uint16            // last error code
+	collation    uint8             // collation used by client, may be different from the collation used by database.
 }
 
 func (cc *clientConn) String() string {