<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

In golang, the compiler performs memory alignment on the fields in the structure.
For example, There are two struct with same field but different order.
```golang
type T1 struct {
	a int8
	b int64
	c int16
}
type T2 struct {
	a int8
	c int16
	b int64
}
```
In `x86_64` platform, each fields will align to `8 byte`, as shown in the figures:
![image](https://user-images.githubusercontent.com/9463871/62467768-19c18a80-b7c7-11e9-9421-4e826c73eabc.png)
Total: 8 byte + 8 byte + 8 byte = 24 byte
![image](https://user-images.githubusercontent.com/9463871/62467803-26de7980-b7c7-11e9-92cf-6d56d4f33403.png)
Total: 8 byte + 8 byte = 16 byte

Different field orders determine the memory size of the structure. 
**Reasonable field orders can reduce memory overhead.**

### What is changed and how it works?

Using [golangci-lint](https://github.com/golangci/golangci-lint) static check all code in TiDB
```
golangci-lint run --disable-all -E maligned
```
And then we deal with the struct needed optimization one by one, there are three cases:
- If the structure is so large that adjusting the field order will greatly affect the readability of the code, no changes will be made.
- If it is an anonymous structure that appears in the `test` file, and it will not affect the memory usage when the program is running, no changes will be made.
- Otherwise, on the premise of modifying the code as little as possible, adjust the field order, make struct take least memory.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - None

Code changes

 - None

Side effects

 - Increased code complexity

Related changes

 - None