@zz-jason 2019-08-05
It's better to add `golangci-lint run --disable-all -E maligned` to the `check-static` section in `Makefile`