@Deardrops 2019-08-08
@zz-jason 
> @Deardrops please add some proper labels for this PR.

Done

> any performance result?

I choose `concatFunction` in `expression/aggregation/concat.go` for the benchmark. source code is [here](https://github.com/pingcap/tidb/pull/11629/files#diff-4f737e39a4674e77e2b98043027e0649R28-R36).

-----
current (48 byte total, reported by `golangci-lint`):
```golang
type concatFunction struct {
	aggFunction
	separator string
	sepInited bool
	maxLen    uint64
	truncated bool
}
```
new (40 byte total):
```golang
type concatFunctionNew struct {
	aggFunction
	separator string
	maxLen    uint64
	sepInited bool
	truncated bool
}
```
-----
Benchmark result:
```
# go test -bench=. -benchmem -benchtime=10s
goos: linux
goarch: amd64
BenchmarkConcatFunction-4       50000000               434 ns/op             280 B/op          0 allocs/op
BenchmarkConcatFunctionNew-4    50000000               331 ns/op             224 B/op          0 allocs/op
PASS
ok      _/root/structure-memory-alignment       39.092s

```

> It's better to add `golangci-lint run --disable-all -E maligned` to the `check-static` section in `Makefile`

In my view, It's unnecessary. Althought all structed will be scanned, not all struct will be changed.(some of them are too complicated to maintain). There is no to add the static check tool into `Makefile`, because it's better for developer to make change by their own eyes, not asked by the tools.