@AilinKid 2019-08-09
The follow-up code should do the check themself, not that automatically.
A quick start is in: GoLand -> Preferences -> Tools -> File Watchers -> Choose golandci-lint Option
when adding code later, it will be triggered automatically.