@zz-jason 2019-08-12
@Deardrops It's better to add `golangci-lint run --disable-all -E maligned` to the `check-static` section in `Makefile`.

Not every programmer knows or realizes the memory alignment of struct. We should use tools to notify them about this. In order to not file such a huge PR in the future.