@Deardrops 2019-08-13
> @Deardrops It's better to add `golangci-lint run --disable-all -E maligned` to the `check-static` section in `Makefile`.
> 
> Not every programmer knows or realizes the memory alignment of struct. We should use tools to notify them about this. In order to not file such a huge PR in the future.

Static check could be the **guide**, but not the **rule**. It's not a good idea that add this static check in `Makefile`.
This PR has optimized 80% struct that need to be optimized, but the rest 20% is maintain.
Everytime programmer run the `golangci-lint`, he / she will get the warning of the rest 20%.
Because of not all struct that need optimized will be optimized, **we should do the static check Manually, leave room for selection.**