@zz-jason 2019-08-13
@Deardrops If we do not require the programmer to check the memory alignment, the percentage of unaligned struct will be increased to more than 20% in the future.

It's hard to make every programmer to check the memory alignment manually every time they file a PR.