@Deardrops 2019-08-13
@zz-jason 
> @Deardrops If we do not require the programmer to check the memory alignment, the percentage of unaligned struct will be increased to more than 20% in the future.
> 
> It's hard to make every programmer to check the memory alignment manually every time they file a PR.

I don't think we should introduce static checking tools into MakeFile.
Some structures are very complex (more than 100 fields), which will affect the readability of the code if optimized. Therefore, we do not need to optimize all structures. Then, static tools scan all structures, including these complex structures that should have been ignored, giving very lengthy prompts. Requiring every programmer who pull request to confirm the tips of static inspection tools, confirm the structures that need optimization or not one by one will increase the workload of merging PR.
Every time PR is filed, it is not necessary to check the memory alignment of the structure, because the results obtained by static tools are too complicated, and manual static checks should be done in PR reviewing and regular checks (such as before the release of new versions).